+++ runway/commands/runway/__init__.py
@@ -2,9 +2,11 @@
 
 # pylint: disable = wildcard-import
 
-from .gen_sample    import *  # noqa
-from .gitclean      import *  # noqa
-from .init          import *  # noqa
-from .preflight     import *  # noqa
-from .test          import *  # noqa
-from .whichenv      import *  # noqa
+from .gen_sample import *  # noqa
+from .gitclean import *  # noqa
+from .init import *  # noqa
+from .preflight import *  # noqa
+from .run_python import *  # noqa
+from .run_stacker import *  # noqa
+from .test import *  # noqa
+from .whichenv import *  # noqa