+++ runway/module/cloudformation.py
@@ -62,6 +62,7 @@ def make_stacker_cmd_string(args, lib_path):
         # Because this will be run via subprocess, the backslashes on Windows
         # will cause command errors
         lib_path = lib_path.replace('\\', '/')
+    # This same code is duplicated in the base runway command `run-stacker`
     return ("import sys;"
             "sys.argv = ['stacker'] + {args};"
             "sys.path.insert(1, '{lib_path}');"
@@ -76,8 +77,40 @@ def make_stacker_cmd_string(args, lib_path):
 class CloudFormation(RunwayModule):
     """CloudFormation (Stacker) Runway Module."""
 
+    def execute_stacker_cmd(self, cmd_list):
+        """Run Stacker in child process."""
+        if getattr(sys, 'frozen', False):
+            # running in pyinstaller single-exe, so sys.executable will
+            # be the all-in-one runway binary
+            executable_cmd_list = [sys.executable, 'run-stacker', '--']
+            LOGGER.debug(
+                "Stacker command being executed: runway-cli %s %s",
+                ' '.join(executable_cmd_list[1:]),
+                ' '.join(cmd_list)
+            )
+            run_module_command(
+                cmd_list=executable_cmd_list + cmd_list,
+                env_vars=self.context.env_vars
+            )
+        else:
+            # traditional python execution
+            stacker_cmd_str = make_stacker_cmd_string(
+                cmd_list,
+                get_embedded_lib_path()
+            )
+            executable_cmd_list = [sys.executable, '-c']
+            LOGGER.debug(
+                "Stacker command being executed: %s \"%s\"",
+                ' '.join(executable_cmd_list),
+                stacker_cmd_str
+            )
+            run_module_command(
+                cmd_list=executable_cmd_list + [stacker_cmd_str],
+                env_vars=self.context.env_vars
+            )
+
     def run_stacker(self, command='diff'):  # pylint: disable=too-many-branches,too-many-locals
-        """Run Stacker."""
+        """Process config files and run Stacker."""
         response = {'skipped_configs': False}
         stacker_cmd = [command, "--region=%s" % self.context.env_region]
 
@@ -141,20 +174,7 @@ class CloudFormation(RunwayModule):
                                         command,
                                         name,
                                         self.context.env_region)
-                            stacker_cmd_str = make_stacker_cmd_string(
-                                stacker_cmd + [name],
-                                get_embedded_lib_path()
-                            )
-                            stacker_cmd_list = [sys.executable, '-c']
-                            LOGGER.debug(
-                                "Stacker command being executed: %s \"%s\"",
-                                ' '.join(stacker_cmd_list),
-                                stacker_cmd_str
-                            )
-                            run_module_command(
-                                cmd_list=stacker_cmd_list + [stacker_cmd_str],
-                                env_vars=self.context.env_vars
-                            )
+                            self.execute_stacker_cmd(stacker_cmd + [name])
                     break  # only need top level files
         return response